+++ 02_testing/diffusion_model.py
@@ -26,3 +26,24 @@ def energy(density, coeff=1.0):
     raise ValueError("Density should be an a *1-dimensional* array of positive integers.")
 
   return sum(density * (density - 1)) * coeff * 0.5
+
+
+
+def energy_aux(density, coeff=2.0):
+  """ A different energy function
+  """
+  # implementation goes here
+
+  # Make sure input is an array
+  density = array(density)
+
+  # of the right kind (integer). Unless it is zero length, in which case type does not matter.
+  if density.dtype.kind != 'i' and len(density) > 0:
+    raise TypeError("Density should be an array of *integers*.")
+  # and the right values (positive or null)
+  if any(density < 0):
+    raise ValueError("Density should be an array of *positive* integers.")
+  if density.ndim != 1:
+    raise ValueError("Density should be an a *1-dimensional* array of positive integers.")
+
+  return sum(density * density) * coeff * 0.5
